requirements.txt
streamlit
transformers
torch
google-api-python-client
google-auth-oauthlib
google-auth-httplib2
pandas
altair
sentencepiece
